A simple C# project
This hospital management system was done by using C# language. In this project, there are two
modules named doctor , diagnosis  and patients. After log in to the system, the user ca check the details of the
doctor and apply for the doctor's appointment and the admin will check the doctor's schedule and
give the appointment. The patient can see the prescription also.